Completed Data layer, Graph, Searching with trie and main activity.

+++ app/src/main/java/com/manish/nalandametro/data/model/Route.kt
@@ -1,7 +1,7 @@
 package com.manish.nalandametro.data.model
 
 data class Route(
-    val id: String?=null,
-    val cost: Int?=null,
-    val distance: Float?=null,
+    val id: String? = null,
+    val cost: Int? = null,
+    val distance: Double? = null,
 )
