+++ app/src/main/java/com/manish/nalandametro/graph/CalculatedPath.kt
@@ -3,7 +3,7 @@ package com.manish.nalandametro.graph
 data class CalculatedPath(
     val type: MetroGraph.PathType,
     val path: List<String>,
-    val distance: Float,
+    val distance: Double,
     val cost: Int,
     val stops: Int
 )
